+++ domain/src/main/java/com/hxl/domain/models/GameResult.kt
@@ -1,3 +1,3 @@
 package com.hxl.domain.models
 
-class GameResult(val difficulty: Int, val levels: Int, val corrects: Int, val time: Double)
+class GameResult(val difficulty: Int, val levels: Int, val corrects: Int, val time: Int)
